# scratch_swagger_new.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "StreamVault",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/Movie/{movieGuid}":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "Movie"
                                                                      ],
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "movieGuid",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK"
                                                                                       }
                                                                           }
                                                         },
                                                 "put":  {
                                                             "tags":  [
                                                                          "Movie"
                                                                      ],
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "movieGuid",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK"
                                                                                       }
                                                                           }
                                                         }
                                             },
...
                  "/api/Review":  {
                                      "post":  {
                                                   "tags":  [
                                                                "Review"
                                                            ],
                                                   "requestBody":  {
                                                                       "content":  {
                                                                                       "application/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/ReviewCreateRequest"
                                                                                                                           }
                                                                                                            },
                                                                                       "text/json":  {
                                                                                                         "schema":  {
                                                                                                                        "$ref":  "#/components/schemas/ReviewCreateRequest"
                                                                                                                    }
                                                                                                     },
                                                                                       "application/*+json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/ReviewCreateRequest"
                                                                                                                             }
                                                                                                              }
                                                                                   }
                                                                   },
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK"
                                                                             }
                                                                 }
                                               }
                                  },
                  "/api/Review/movie/{movieGuid}":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "Review"
                                                                             ],
                                                                    "parameters":  [
                                                                                       {
                                                                                           "name":  "movieGuid",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "string"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "OK"
                                                                                              }
                                                                                  }
                                                                }
                                                    },
                  "/api/Role":  {
                                    "post":  {
                                                 "tags":  [
                                                              "Role"
                                                          ],
                                                 "requestBody":  {
                                                                     "content":  {
                                                                                     "application/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/CreateRoleDto"
                                                                                                                         }
                                                                                                          },
                                                                                     "text/json":  {
                                                                                                       "schema":  {
                                                                                                                      "$ref":  "#/components/schemas/CreateRoleDto"
                                                                                                                  }
                                                                                                   },
                                                                                     "application/*+json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/CreateRoleDto"
                                                                                                                           }
                                                                                                            }
                                                                                 }
                                                                 },
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "OK"
                                                                           }
                                                               }
                                             },
                                    "get":  {
                                                "tags":  [
                                                             "Role"
                                                         ],
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "OK"
                                                                          }
                                                              }
                                            }
                                },
                  "/api/Role/{roleGuid}":  {
                                               "put":  {
                                                           "tags":  [
                                                                        "Role"
                                                                    ],
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "roleGuid",
                                                                                  "in":  "path",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "requestBody":  {
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/UpdateRoleDto"
                                                                                                                                   }
                                                                                                                    },
                                                                                               "text/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/UpdateRoleDto"
                                                                                                                            }
                                                                                                             },
                                                                                               "application/*+json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/UpdateRoleDto"
                                                                                                                                     }
                                                                                                                      }
                                                                                           }
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK"
                                                                                     }
                                                                         }
                                                       },
                                               "delete":  {
                                                              "tags":  [
                                                                           "Role"
                                                                       ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "roleGuid",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "string"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK"
                                                                                        }
                                                                            }
                                                          }
                                           },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateRoleDto":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "name":  {
                                                                                             "type":  "string",
                                                                                             "nullable":  true
                                                                                         }
                                                                            },
                                                             "additionalProperties":  false
                                                         },
...
                                       "ReviewCreateRequest":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "movieGuid":  {
                                                                                                        "type":  "string",
                                                                                                        "nullable":  true
                                                                                                    },
                                                                                      "rating":  {
                                                                                                     "type":  "integer",
                                                                                                     "format":  "int32"
                                                                                                 },
                                                                                      "reviewText":  {
                                                                                                         "type":  "string",
                                                                                                         "nullable":  true
                                                                                                     }
                                                                                  },
                                                                   "additionalProperties":  false
                                                               },
...
                                       "UpdateRoleDto":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "name":  {
                                                                                             "type":  "string",
                                                                                             "nullable":  true
                                                                                         }
                                                                            },
                                                             "additionalProperties":  false
                                                         },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "apiKey",
                                                              "description":  "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter \u0027Bearer\u0027 [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\"",
                                                              "name":  "Authorization",
                                                              "in":  "header"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ],
...
}